backend/requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
aiosqlite==0.17.0
anyio==3.4.0
asgiref==3.4.1
async-timeout==4.0.1
asyncpg==0.25.0
attrs==21.2.0
bcrypt==3.2.0
cffi==1.15.0
charset-normalizer==2.0.9
click==8.0.3
dnspython==2.1.0
email-validator==1.1.3
fastapi==0.70.1
fastapi-users==8.1.3
fastapi-users-db-tortoise==1.0.0
frozenlist==1.2.0
h11==0.12.0
idna==3.3
iso8601==0.1.16
makefun==1.12.1
multidict==5.2.0
passlib==1.7.4
pycparser==2.21
pydantic==1.8.2
PyJWT==2.3.0
pypika-tortoise==0.1.1
python-multipart==0.0.5
pytz==2021.3
six==1.16.0
sniffio==1.2.0
starlette==0.16.0
tortoise-orm==0.17.8
typing-extensions==4.0.1
uvicorn==0.16.0
yarl==1.7.2
